trip_logger/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.4
django-cors-headers==4.2.0
whitenoise==6.5.0
gunicorn==21.2.0